+++ Simple Building Simulator/occupancy.h
@@ -12,7 +12,7 @@ namespace SimpleBuildingSimulator {
 	public:
 		Occupants();
 		~Occupants();
-		Eigen::MatrixXi GetOccupancyForecast(uint32 duration, uint16 time_step, uint8 num_zones, uint8 num_rooms);
+		Eigen::MatrixXi GetOccupancyForecast(long int duration, int time_step, int num_zones, int num_rooms);
 	private:
 
 	};
